requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
Flask==2.3.3
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
groq==0.31.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.3.0
numpy==2.0.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==3.20.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2025.2
pywin32==311
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.12.0
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
xlsxwriter==3.2.5
zipp==3.23.0
